requirements.txt
python-telegram-bot>=20.0
discord.py==2.3.2
python-dotenv==1.0.0
schedule==1.2.1
psutil>=5.9.0 
aiohttp>=3.7.4